+++ Assets/Scripts/BuildJobBurstParallel.cs
@@ -13,7 +13,7 @@ using UnityEngine;
         public int3 dimension;
         public int groundHeight;
         public int terrainHeight;
-        public NativeArray<float> map;
+        public NativeArray<int> map;
         public TerrainSettings Settings;
 
         public void Execute(int index)
